requirements.txt
streamlit
pillow
opencv-python
numpy
pytesseract
ultralytics
easyocr
